requirements.txt
# Gwen Voice Assistant - Complete Dependencies

# === Core Wake Word Detection ===
pvporcupine>=3.0.0

# === Audio Processing ===
pyaudio>=0.2.11

# === Speech Recognition ===
SpeechRecognition>=3.10.0

# === Text-to-Speech ===
elevenlabs>=0.2.26
pyttsx3>=2.90  # Fallback TTS

# === Audio Playback ===
playsound>=1.3.0

# === Windows Integration ===
pywin32>=306; sys_platform == "win32"
winshell>=0.6; sys_platform == "win32"

# === Environment Management ===
python-dotenv>=1.0.0

# === HTTP Requests ===
requests>=2.31.0

# === File Operations ===
pathlib2>=2.3.7; python_version < "3.4"

# === Development Dependencies (Optional) ===
# Uncomment these for development
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0

# === Installation Notes ===
# For Windows users:
# - Install Microsoft C++ Build Tools if pyaudio installation fails
# - Some packages may require Visual Studio Build Tools
#
# For macOS users:
# - Install portaudio: brew install portaudio
# - May need Xcode command line tools
#
# For Linux users:
# - Install portaudio: sudo apt-get install portaudio19-dev
# - Install python3-dev: sudo apt-get install python3-dev